requirements.txt
aniso8601==9.0.1
attrs==21.2.0
click==8.0.3
coverage==6.2
Flask==2.0.2
Flask-JWT-Extended==4.3.1
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.26.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.11.0
pycryptodomex==3.11.0
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.5
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.26
toml==0.10.2
Werkzeug==2.0.2
